Fix Report showing numbered names
When the HarmonyMod is unloaded last, before all other mods are unloaded,
their Name and Description properties are not accessible, and their
userInstance reference is also null.

Mods that had assemblies post unload (not clear how this happens, yet,
but I suspect it is related to throwing an exception during onDisable())
the report would attempt to read their name and throw a NullReference
because userModInstance was null

Mods that did not have assemblies post unload would be shown as just their
workshop item number.

This commit fixes #2 and #4 by caching the Name and Description fields of
all mods just after all mods are loaded (on m_introLoaded event)