Implement Late Report
Previously, the Final Harmony Report would be written to the log file
when the mod was disabled. If it was not the last mod to be disabled on
shutdown, this report would miss any faults that occured after it was
written.

This commit hooks PluginManager.OnDestroy() in order to write the Final
Report after the last mod has been unloaded.