Implement ACL
User mods should not be able to remove Harmony's patches, UnpatchAll(),
or patch the harmony libs or the mod itself. This would destabilize the
patching environment, and indeed is what happens as some mods like
PurchaseIt! call UnpatchAll() in their OnLevelUnload() handler. This
breaks the mod environemnt when the game returns to main menu.

In the future, the ACL mechanism can be expanded to deny mods trying to
patch key game methods.

This commit implements a basic ACL mechanism, where patcher, patched
method and patch are checked when a Patch() or Unpatch() request is made.

Implements issue #10