Catch ReflectionTypeLoadException from mods
Mods that are missing dependencies throw the type load exception (TLE) when
instantiated or probed by fetching PluginInfo.userModInstance, but the
resulting stack trace does not attribute the fault to the mod that caused
it.

In order to track proper attribution, this commit installs an alternate
getter for userModInstance, which reports the offending mod directly
when the TLEs are thrown on construction.

A mod can still throw TLE during operation, eg, if it attempts to use
a class from another assembly that is missing, but that TLE's stack trace
will contain the loading mod, and thus will be attributed to it.

The Harmony Mod did not handle mods broken this way.

This commit fixes the handling of other mods being unable to load due to
this bug, catches those TLE exceptions, and adds them to the Harmony Report

Fixes issue #9