Implement API Misuse detection
Mods that don't implement the API calls to check in Harmony is installed
and to schedule patching when ready cause spurious failures in some cases:

- when harmony is not yet installed when they start
- when a signed library is present while mod was built with unsigned

Additionally, when a Harmony member variable exists in a public type
(ExportedType), it must be resolved when the PluginManager AddAssembly(),
after having enumerated all present assemblies and noted their paths.

Together, this caused the signed 0Harmony.dll shipped with the mod to not
be a match for a mod's reference. That mod was loaded before Harmony, and
had a public static class containing a Harmony member. This combination
of failures lead to "TypeResolveException" when the game tried to
AddAssembly the assembly with unresolved dependency.

If the public class was internal instead, it would have been sufficient
to not cause the TypeResolveException, but the Resolve failures when
Harmony is not yet installed would still happen, of course.

This commit detects when mods do not check in Harmony is installed first,
and warns in the log file (CitiesHarmony.API misuse ... will be obsolete
at v1.1 release)

Fixes issue #8