WIP Fix Option button opening Report
In the Content Manager, for mods near the Harmony Mod, which had an OPTIONS
button, clicking the OPTIONS button would open the Report instead.

The cause is that the event handlers were not reset correctly for mods
which are not Harmony.

This commit resets the eventClick at the same time as reseting the button
text.

This fixes Issue #5