Handle Felix Schmidt 03-29-2021 DMCA complaint
In response to Felix Schmidt's complaint to GitHub, I am adding his
copyright notice in code fragments that contain fragments of his code.

Although he did not post his own notices on the code, he demands that I
include it. While his complaint is fraudulent in my opinion, and will be
pursuing it as such, this commit should satisfy GitHub so that they will
not remove my repository as would otherwise be required by DMCA law.

All the code here is quite low quality, and it is in part responsible for
many various systematic breakages in the Cities Skylines mod collection
on Colossal Order's workshop.

The API, ExampleMod and README.md files are not used at all in the
compiled product, because they are just bad.

I have copied Felix Schmidt's Harmony1SelfPatcher and StateTransfer
because although they are quite poor, I considered they would do no more
harm than his own project does, and because it takes some time to
re-engineer them.

The TypeExtensions.cs is required by his self patcher, and will be removed
as soon as the Harmony1 compatibility is re-engineered, as it is not
necessary.

The Tests relating to structs and specials are marginally useful.

The README.md gives particularly bad examples on how to use Harmony in a
mod. A complete re-write is coming soon.

Finally, as Felix Schmidt's made a complaint agaist all fragments of his
code, it is reasonable to assume that if you use any of his example code
in your mod, you are potentially liable to a similar DMCA complaint against
your mod. Buyer beware.

The GPL License terms are amended to exlude Felix Schmidt and his employer,
Colossal Order, from copying this code.