Change API's namespace to Harmony
The API namespace used to be CitiesHarmony.API.

The name is specifically chosen to conflict with HarmonyLib.Harmony.
Mods should either use the Harmony or the HarmonyLib namespace in any one
type, but not both. If both are used, additional disambiguation qualifiers
are needed, which should be an early warning that something is wrong.

# Conflicts:
#	ExampleMod/ExampleMod.cs

fixup AssemblyInfo.cs

fixup API/HarmonyAPIException.cs