Add Test.ExploitCatch for issue #18
This test demonstrates reliance on the resolver vulnerability
reported as issue #18, and how this reliance can break the mod.

The test references VersionedAssembly[1.0.0.0], but ships
with VersionedAssembly[2.0.0.0], which is not (fully) backwards
compatible. A feature is removed from 2.0.0.0, which will
cause the test to fail due to being unable to find the feature it
expects (which was present at 1.0.0.0, which the test was built
for).

The resolver offers the only version it has available, 2.0.0.0,
when 1.0.0.0 is requested, assuming that all assemblies that are
"similar" are equivalent, regardless of version.

This test demonstrates how the resolver fails, and also how the
mod fails.