Disable the malware
- remove adware for DLC
- remove adware for workshop
- block exfiltration of data to "Paradox Online Publishing Services"
- create error message for mods abusing the assembly resolver vulnerability

## Keylogger

There is **keylogging functionality in the released vanilla game**,
which sends data out to "Paradox Online Publishing Services". It is
contained in:

steamapps\common\Cities_Skylines\Cities_Data\Plugins\pops_api.dll
steamapps\common\Cities_Skylines\Cities_Data\Managed\PopsApiWrapper.dll

## Credentials transmitted insecurely

When a player logs into Paradox, **login information is transmitted insecurely
over the internet**, from Starter.cs:

ParadoxAccountCreate.landingUrl = "http://www.citiesskylines.com/auth/cas"

The credentials are stored in the clear in the gameSettings.cgs file, and
loaded ParadoxAccountPanel.cs; there is no modern session-token based
account security implemented.

I haven't verified it, but this programming pattern of insecure network
resources (http, not https), combined with storage of password in the
clear, indicates that little or no consideration was given the the user's
credentials security.

## Granting Paradox full access to user's Steam account

The Paradox login procedure can also use a linked steam account,
specifically the authentication token issued to the end-user
by Steam. It appears that this token (which belongs to the
end user), is transmitted to Paradox:

private IEnumerator LoginWithSteamTokenCoroutine(Action<PopsManager.
AccountActionResult> callback)

If correct, this is a **huge lapse in security** as it allows
Paradox to access the user's Steam account with the user's security
clearance (ie, Paradox would be able to access all of the user's
information that the user would be able to access).

I have not analyzed in detail what would be sent, as for the purposes
of the Harmony Mod, it is sufficient to disable the entire pops_api
function to eliminate these risks.

Blocking the Paradox login prevents sensitive login credentials from
being transmitted in the clear over the internet, which itself would
be a security hazard.

## Assembly resolution vulnerability.

The game's implementation of the assembly name resolver is insecure,
and allows man-in-the-middle attacks, as well as hijacking control
by an unscrupulous mod author.

In the function CurrentDomain_AssemblyResolve(), whenever an
incorrectly coded mod requires resolution, a match is naively
searched, without regard to version numbers, and based on partial
name match. This means that if some functionality requires for instance
the services of a StrongNamed assembly named "TransferBroker", it
can be MITM'd with a malicious impostor assembly named "T" or "Tr"
or any such matching name.

The offending code is the partial match:

AssemblyName name = assembly2.GetName();
if (args.Name.StartsWith(name.Name))
{
	CODebugBase<LogChannel>.VerboseWarn(LogChannel.Serialization, string.Concat(new string[]
	{
		"Assembly '",
		args.Name,
		"' resolved to '",
		assembly2.Location,
		"'"
	}));
	assembly = assembly2;
	break;
}
}

Many mods inadvertently use this loose match vulnerability by
being compiled wrong, such as @kian.zarrin's code, which
includes many references to assembly version that don't exist, ie,
"missing TrafficManager[1.0.0.0]" type of errors from the Harmony
Report.

Since many mods rely on this broken behaviour, i have emulated
it in the function
Behave_Like_ColossalOrder_Assembly_Resolver_Vulnerability_YUK_YUK
_YUK_WHY_DO_I_FEEL_DIRTY_NOW()

Fixing it would break many mods that rely on this bug to be in
place. Hopefully the function name will be noticed and the bug
properly fixed, eventually.

Colossal Order has completely cut off communications with me
and I have no means to alert them of the vulnerabilities
to which they subject their end users.

Includes detection and workaround for #18 (exploiting game's resolver
vulnerability)

This commit eliminates the adware windows at the main screen,
blocks the Paradox Online Publishing Services to send
the user's credentials insecurely, to transmit the user's
Steam authentication token to Paradox, and adds a new error
message for mods susceptible to the AssemblyResolver
vulnerability. The error message will refer users to
issue #18