Reformat Report to show authorName
The author name is more useful to the end user, and will be shown
on RELEASE builds. DEBUG builds will continue to show the harmony version
and the API version.

Shrink field widths to make the table easier to format in forum posts.
It's still wider than Steam's forum, even after 14-char shrink of each line.