Add Mistarget test
This mod is targetted to .Net 4 instead of the correct v3.5
This would cause "[err] missing" lines on the Report, but usually
nothing "bad" happens.

This test demonstrates why this seemingly benign error can be
a real error in some circumstances.

This test relies on a feature of .Net 4 that works differently
than it did in v3.5. Therefore a mod built for .net 4 would
misbehave if it didn't properly account for the difference.